[Refactor]: modify deleteWallet
- ㅇepending on the value of the parameter, package information is not deleted.

+++ source/did-wallet-sdk-aos/src/main/java/org/omnione/did/sdk/core/api/WalletService.java
@@ -274,7 +274,9 @@ public class WalletService implements WalletServiceInterface {
             @Override
             public void run() {
                 DBManager walletDB = DBManager.getDatabases(context);
-                walletDB.caPkgDao().deleteAll();
+                if (deleteAll) {
+                    walletDB.caPkgDao().deleteAll();
+                }
                 walletDB.userDao().deleteAll();
                 walletDB.tokenDao().deleteAll();
             }
